Using [pino](https://github.com/mcollina/pino)'s own benchmark suite to add some perf improvements .. some are kind of crazy (there's a good reason I made a separate file called format.js). Mostly it's not down to the speed of pino, but it's at least competitive now.

API hasn't really changed, except now it's limited to 16 arguments on a call to a log function, I think this is more than reasonable but still it's technically a breaking change. There's also a new `bole.setFastTime()` function to switch to `Date.now()` which is the pino default (personally I hate logs that use epoch time but it is quite fast so we can give the option at least).

It's also very possible that there is accidental breakage in the API here given all of the shuffling of logic so bumping semver-major if we go ahead with this is probably wise regardless.

Below are benchmarks comparing old and new bole as well as bole with `fastTime` enabled. The benchmarks were run with: ```for i in `ls ../pino/benchmarks/`; do echo $i; node ../pino/benchmarks/$i ; done``` having checked pino out, done an `npm install` and linked bole into pino's node_modules. The `fastTime` ones were done by adding `.setFastTime(true)` to all of the bole setup sections in the files.

Happy to hear of more crazy ways we can speed this up! I'm not sold on pino's "extreme" mode fwiw, too much hairy logic to get that right.

/cc @yoshuawuyts 


```
basic.js
benchBunyan*10000:     1608.165ms
benchWinston*10000:    2582.654ms
benchOldBole*10000:    1774.625ms
benchNewBole*10000:    585.776ms
benchNewestBole*10000: 372.713ms
benchDebug*10000:      545.641ms
benchLogLevel*10000:   437.950ms
benchPino*10000:       370.170ms
benchPinoExreme*10000: 142.827ms
benchBunyan*10000:     1627.781ms
benchWinston*10000:    2638.942ms
benchOldBole*10000:    1637.006ms
benchNewBole*10000:    574.749ms
benchNewestBole*10000: 352.200ms
benchDebug*10000:      527.688ms
benchLogLevel*10000:   394.412ms
benchPino*10000:       324.713ms
benchPinoExreme*10000: 134.498ms


child.js
benchBunyanChild*10000:      1717.793ms
benchOldBoleChild*10000:     1625.616ms
benchNewBoleChild*10000:     616.279ms
benchNewestBoleChild*10000:  458.978ms
benchPinoChild*10000:        421.797ms
benchPinoExtremeChild*10000: 190.450ms
benchBunyanChild*10000:      1735.584ms
benchNewBoleChild*10000:     596.083ms
benchNewestBoleChild*10000:  439.364ms
benchPinoChild*10000:        405.750ms
benchPinoExtremeChild*10000: 174.887ms


childCreation.js
benchBunyanCreation*10000:      1912.047ms
benchNewBoleCreation*10000:     674.235ms
benchNewestBoleCreation*10000:  473.900ms
benchPinoCreation*10000:        455.381ms
benchPinoExtremeCreation*10000: 222.714ms
benchBunyanCreation*10000:      1958.595ms
benchOldBoleChild*10000:        1597.697ms
benchNewBoleCreation*10000:     636.196ms
benchNewestBoleCreation*10000:  443.747ms
benchPinoCreation*10000:        454.718ms
benchPinoExtremeCreation*10000: 204.472ms


deep-object.js
benchBunyanDeepObj*10000:            9922.489ms
benchWinstonDeepObj*10000:           20467.637ms
benchOldBoleDeepObj*10000:           22942.288ms
benchNewBoleDeepObj*10000:           5334.259ms
benchNewestBoleDeepObj*10000:        5976.224ms
benchLogLevelDeepObj*10000:          51442.827ms
benchPinoDeepObj*10000:              5829.747ms
benchPinoUnsafeDeepObj*10000:        4346.752ms
benchPinoExtremeDeepObj*10000:       5941.517ms
benchPinoUnsafeExtremeDeepObj*10000: 4320.007ms
benchBunyanDeepObj*10000:            10368.592ms
benchWinstonDeepObj*10000:           20480.133ms
benchOldBoleDeepObj*10000:           23617.181ms
benchNewBoleDeepObj*10000:           5171.666ms
benchNewestBoleDeepObj*10000:        5913.482ms
benchLogLevelDeepObj*10000:          51737.796ms
benchPinoDeepObj*10000:              5708.446ms
benchPinoUnsafeDeepObj*10000:        4365.840ms
benchPinoExtremeDeepObj*10000:       5984.172ms
benchPinoUnsafeExtremeDeepObj*10000: 4188.128ms


multiArg.js
benchBunyanMulti*10000:                       1515.135ms
benchWinstonMulti*10000:                      2647.043ms
benchOldBoleMulti*10000:                      1856.140ms
benchNewBoleMulti*10000:                      594.838ms
benchNewestBoleMulti*10000:                   377.063ms
benchLogLevelMulti*10000:                     444.528ms
benchDebugMulti*10000:                        516.222ms
benchPinoMulti*10000:                         360.566ms
benchPinoExtremeMulti*10000:                  150.123ms
benchDebugInterpolate*10000:                  665.639ms
benchBunyanInterpolate*10000:                 1533.101ms
benchWinstonInterpolate*10000:                2809.965ms
benchOldBoleInterpolate*10000:                1845.031ms
benchNewBoleInterpolate*10000:                588.014ms
benchNewestBoleInterpolate*10000:             366.402ms
benchPinoInterpolate*10000:                   491.378ms
benchPinoExtremeInterpolate*10000:            192.419ms
benchBunyanInterpolateAll*10000:              1780.023ms
benchDebugInterpolateAll*10000:               803.718ms
benchWinstonInterpolateAll*10000:             2956.432ms
benchOldBoleInterpolateAll*10000:             2055.106ms
benchNewBoleInterpolateAll*10000:             686.849ms
benchNewestBoleInterpolateAll*10000:          454.530ms
benchLogLevelInterpolateAll*10000:            623.155ms
benchPinoInterpolateAll*10000:                547.891ms
benchPinoExtremeInterpolateAll*10000:         303.352ms
benchPinoUnsafeInterpolateAll*10000:          565.622ms
benchPinoUnsafeExtremeInterpolateAll*10000:   309.048ms
benchDebugInterpolateExtra*10000:             807.428ms
benchBunyanInterpolateExtra*10000:            3179.967ms
benchWinstonInterpolateExtra*10000:           3169.942ms
benchOldBoleInterpolateExtra*10000:           3736.959ms
benchNewBoleInterpolateExtra*10000:           1926.395ms
benchNewestBoleInterpolateExtra*10000:        1548.289ms
benchLogLevelInterpolateExtra*10000:          1759.567ms
benchPinoInterpolateExtra*10000:              620.915ms
benchPinoUnsafeInterpolateExtra*10000:        628.565ms
benchPinoExtremeInterpolateExtra*10000:       397.313ms
benchPinoUnsafeExtremeInterpolateExtra*10000: 375.254ms
benchDebugInterpolateDeep*10000:              663.052ms
benchBunyanInterpolateDeep*10000:             6041.667ms
benchWinstonInterpolateDeep*10000:            22134.751ms
benchOldBoleInterpolateDeep*10000:            6587.924ms
benchNewBoleInterpolateDeep*10000:            4741.358ms
benchNewestBoleInterpolateDeep*10000:         4217.562ms
benchLogLevelInterpolateDeep*10000:           3323.441ms
benchPinoInterpolateDeep*10000:               6017.806ms
benchPinoUnsafeInterpolateDeep*10000:         4708.956ms
benchPinoExtremeInterpolateDeep*10000:        6533.090ms
benchPinoUnsafeExtremeInterpolateDeep*10000:  4726.833ms
benchBunyanMulti*10000:                       1556.186ms
benchWinstonMulti*10000:                      2641.214ms
benchOldBoleMulti*10000:                      1873.938ms
benchNewBoleMulti*10000:                      583.086ms
benchNewestBoleMulti*10000:                   355.907ms
benchLogLevelMulti*10000:                     470.015ms
benchDebugMulti*10000:                        518.127ms
benchPinoMulti*10000:                         358.284ms
benchPinoExtremeMulti*10000:                  150.430ms
benchDebugInterpolate*10000:                  666.039ms
benchBunyanInterpolate*10000:                 1544.669ms
benchWinstonInterpolate*10000:                2633.577ms
benchOldBoleInterpolate*10000:                1803.603ms
benchNewBoleInterpolate*10000:                579.716ms
benchNewestBoleInterpolate*10000:             368.567ms
benchPinoInterpolate*10000:                   400.438ms
benchPinoExtremeInterpolate*10000:            179.811ms
benchBunyanInterpolateAll*10000:              1734.928ms
benchDebugInterpolateAll*10000:               830.598ms
benchWinstonInterpolateAll*10000:             2914.816ms
benchOldBoleInterpolateAll*10000:             2108.744ms
benchNewBoleInterpolateAll*10000:             668.581ms
benchNewestBoleInterpolateAll*10000:          459.191ms
benchLogLevelInterpolateAll*10000:            564.736ms
benchPinoInterpolateAll*10000:                566.521ms
benchPinoExtremeInterpolateAll*10000:         314.738ms
benchPinoUnsafeInterpolateAll*10000:          552.270ms
benchPinoUnsafeExtremeInterpolateAll*10000:   299.188ms
benchDebugInterpolateExtra*10000:             805.610ms
benchBunyanInterpolateExtra*10000:            3229.008ms
benchWinstonInterpolateExtra*10000:           3169.397ms
benchOldBoleInterpolateExtra*10000:           3462.622ms
benchNewBoleInterpolateExtra*10000:           1865.117ms
benchNewestBoleInterpolateExtra*10000:        1549.994ms
benchLogLevelInterpolateExtra*10000:          1897.569ms
benchPinoInterpolateExtra*10000:              642.484ms
benchPinoUnsafeInterpolateExtra*10000:        623.055ms
benchPinoExtremeInterpolateExtra*10000:       394.064ms
benchPinoUnsafeExtremeInterpolateExtra*10000: 365.378ms
benchDebugInterpolateDeep*10000:              667.786ms
benchBunyanInterpolateDeep*10000:             6059.128ms
benchWinstonInterpolateDeep*10000:            22247.574ms
benchOldBoleInterpolateDeep*10000:            6609.189ms
benchNewBoleInterpolateDeep*10000:            4562.093ms
benchNewestBoleInterpolateDeep*10000:         4262.331ms
benchLogLevelInterpolateDeep*10000:           3293.833ms
benchPinoInterpolateDeep*10000:               6101.653ms
benchPinoUnsafeInterpolateDeep*10000:         4854.862ms
benchPinoExtremeInterpolateDeep*10000:        7011.338ms
benchPinoUnsafeExtremeInterpolateDeep*10000:  4695.801ms


object.js
benchBunyanObj*10000:                         1678.507ms
benchWinstonObj*10000:                        2656.642ms
benchOldBoleObj*10000:                        1588.992ms
benchNewBoleObj*10000:                        597.381ms
benchNewestBoleObj*10000:                     415.290ms
benchLogLevelObject*10000:                    1862.047ms
benchPinoObj*10000:                           398.471ms
benchPinoUnsafeObj*10000:                     461.343ms
benchPinoExtremeObj*10000:                    205.798ms
benchPinoUnsafeExtremeObj*10000:              191.235ms
benchBunyanObj*10000:                         1687.752ms
benchWinstonObj*10000:                        2791.879ms
benchOldBoleObj*10000:                        1562.442ms
benchNewBoleObj*10000:                        586.917ms
benchNewestBoleObj*10000:                     380.834ms
benchLogLevelObject*10000:                    1945.977ms
benchPinoObj*10000:                           368.611ms
benchPinoUnsafeObj*10000:                     476.259ms
benchPinoExtremeObj*10000:                    238.516ms
benchPinoUnsafeExtremeObj*10000:              200.529ms
```