@rvagg 2016-06-17
fwiw I tried to do some manual JSON strinficiation, creating the initial part of the object and tacking on a stringification of the rest of the object if there is any, but it resulted in a slowdown, even for the simple case where you don't need to add any additional stringification to the manual bit. I also tried out pino's "faster" util.format but it didn't help at all.

I'm doing this all on Node 6, perhaps it's different on older versions.