@brycebaril 2016-06-17
I made a bole-only version of one of the benchmarks: [multiArg.js](https://dl.dropboxusercontent.com/u/6789610/bole/multiArg.js)

Here's the [v8-profiler output](https://dl.dropboxusercontent.com/u/6789610/bole/bench.cpuprofile)

The majority of the time is spent in bole's `namedLevelLogger` after that it's mostly formatting.

Here's the IRHydra artifacts: [code.asm](https://dl.dropboxusercontent.com/u/6789610/bole/code.asm) and [hydrogen-30454-1.cfg](https://dl.dropboxusercontent.com/u/6789610/bole/hydrogen-30454-1.cfg)

IRHydra doesn't show any deopts in `namedLevelLogger` or any other part of `bole` but it does show lots of deopts in core. (timers, buffer, and streams)

There were some deopts in `util.format` and associated code, so bypassing that entirely could be one approach, though based on the v8-profiler output, something to remove instruction count in `namedLevelLogger` would be where I'd start.