@brycebaril 2016-06-17
I also generated a flamegraph you can find here: http://brycebaril.com/profile-32255/flamegraph.html

Definitely suggests looking at ways to replace or work better with `util.format` but there are some other potential Buffer integrations to consider. Also JSON.stringify is a hit.