@yoshuawuyts 2016-06-17
`pino` uses [quick-format](https://www.npmjs.com/package/quick-format), it might perhaps work as a proper substitution for `util.format` here