@rvagg 2016-06-18
I did originally try quick-format but it _added_ time rather than making it any quicker, fast-safe-stringify is in this PR and fast-json-parse isn't necessary for bole.