@rvagg 2016-06-18
OK folks, I've _mostly_ cracked it. It seems that the bulk of the time is spent in `stringify()`, I was thinking that a lot of the cost was in `write()`, so I revisited the partial manual stringification and did it a tiny bit differently than before, plus I added a bunch of other micro optimisations in here and I've updated the benchmarks in the OP with new numbers for `boleNewest` (also updated the Pino numbers so they are from the same run. You can see that we are within sneezing distance for the basic benchmark. I don't know where else I can shave off time and am not really sure what the edge is that pino has left here.

Interestingly, if you look at the interpolation tests, we're doing very well, and in some we're beating pino by a non-trivial amount. I think the cost for pino here is in `quick-format` because if I switch out at this stage, bole slows down by about that much.

The object benchmarks show a roughly similar difference between bole and pino as the basic ones, since objectMode bypasses the stringification completely, I believe this says that I've managed to completely remove the difference between them in stringification (except for deeper object I suspect). So there's something left here!

If nothing else comes of this I'm going to publish as a semver-major and submit a PR to pino with an update to the version & bole setup in their benchmarks for fastTime and suggest they change the language to "pino is _one of_ the fastest logging libraries", just for fun!