@rvagg 2016-06-18
weird, I believe I've just reduced the difference in the basic.js benchmark to be statistically insignificant _with this one weird trick_:

```
diff --git a/bole.js b/bole.js
index 015d5f8..d76ae67 100644
--- a/bole.js
+++ b/bole.js
@@ -88,7 +88,11 @@ function stringify (level, name, message, obj) {
   for (k in obj)
     s += ',' + _stringify(k) + ':' + _stringify(obj[k])

-  return s + '}'
+  s += '}'
+
+  Number(s) // convert internal representation to plain string
+
+  return s
 }
```

Thanks to the hint from https://github.com/davidmarkclements/flatstr which pino uses. Have a read of the README there, I didn't know this about concatenated strings.