@rvagg 2016-06-18
then .. applying the same trick to a pre-cached section of the stringified output (hostname, pid, level) and using that instead of the extra concatenation when performing stringification .. gets us to a statistically significant winning position in the basic! pino would have difficulty with this because it allows you to change the hostname and its output is jumbled up with the potentially cachable bits not next to eachother so it'd be a breaking change (sort of) if they were to do it with the tiny bit that they could do it with.