@rvagg 2016-06-19
merged and released as v3, but the challenge still exists if anyone wants to try and squeeze more out of this